fusedump: read packets lazily from the stream
Previously the packets were all read into a list which required a
lot of memory for long traces. This patch reads the packets lazily
from the stream, just in time to pretty-print them. This reduces the
memory footprint of fusedump and removes the initial start-up delay
where previously the whole file was parsed.

Signed-off-by: David Scott <dave.scott@docker.com>